feat: adding skeleton loading in task detail page
